@@ -9,17 +9,18 @@
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.1",
     "@netlify/plugin-nextjs": "^5.9.4",
+    "@testing-library/dom": "^10.4.0",
     "dotenv": "^16.4.5",
-    "googleapis": "^109.0.1",
+    "googleapis": "^144.0.0",
     "moment": "^2.29.4",
-    "next": "^14.2.3",
+    "next": "^15.1.6",
     "next-seo": "^6.5.0",
     "react": "^18.3.1",
     "react-accessible-accordion": "^5.0.0",
     "react-big-calendar": "^1.8.4",
     "react-countdown": "^2.3.5",
     "react-dom": "^18.3.1",
-    "sass": "^1.58.3",
+    "sass": "^1.84.0",
     "yarn": "^1.22.21"
   },
   "scripts": {
@@ -68,7 +69,7 @@
   "devDependencies": {
     "@axe-core/react": "^4.7.3",
     "@testing-library/jest-dom": "^6.1.3",
-    "@testing-library/react": "^15.0.7",
+    "@testing-library/react": "^16.2.0",
     "csv-parser": "^3.0.0",
     "eslint": "^7.32.0",
     "eslint-config-next": "^13.4.2",
@@ -77,12 +78,12 @@
     "eslint-plugin-react": "^7.33.2",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
-    "jest-axe": "^7.0.0",
+    "jest-axe": "^9.0.0",
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.4.31",
-    "stylelint": "^13.13.1",
-    "stylelint-config-recommended": "^5.0.0",
-    "stylelint-config-sass-guidelines": "^8.0.0",
-    "stylelint-scss": "^3.21.0"
+    "stylelint": "^16.14.1",
+    "stylelint-config-recommended": "^15.0.0",
+    "stylelint-config-sass-guidelines": "^12.1.0",
+    "stylelint-scss": "^6.11.0"
   }
 }
