@@ -1,4 +1,4 @@
-import React from 'react';
+import { useState, useEffect } from 'react';
 
 // this can get refactored into something better
 const mapUpToSum = (num, fn) => {
@@ -30,18 +30,16 @@ const generateCols = (n, m, classPattern = [''], randomize = false) => {
   ));
 };
 
-export default class Banner extends React.Component {
-  constructor(props) {
-    super(props);
-    this.timer = null;
-    this.color = 0;
-    this.state = { randomize: false };
-  }
+const Banner = (props) => {
+  const [randomize, setRandomize] = useState(false);
+  const [color, setColor] = useState(0);
+  let timer;
 
-  componentDidMount() {
-    this.setState({ randomize: true });
+  useEffect(() => {
+    setRandomize(true);
+
     const committees = ['acm'];
-    if (!this.props.decorative)
+    if (!props.decorative) {
       committees.push(
         'studio',
         'icpc',
@@ -52,78 +50,78 @@ export default class Banner extends React.Component {
         'ai',
         'hack',
       );
+    }
+
     const el = document.querySelector('.banner');
-    this.timer = setInterval(() => {
-      el.classList.remove(committees[this.color]);
-      this.color = (this.color + 1) % committees.length;
-      el.classList.add(committees[this.color]);
-      this.forceUpdate();
+    timer = setInterval(() => {
+      el.classList.remove(committees[color]);
+      setColor((prevColor) => (prevColor + 1) % committees.length);
+      el.classList.add(committees[color]);
     }, 4000);
-  }
 
-  componentWillUnmount() {
-    clearInterval(this.timer);
-    this.timer = null;
-  }
+    return () => {
+      clearInterval(timer); // Cleanup on unmount
+    };
+  }, [color, props.decorative]); // Re-run when color or decorative props change
 
-  render() {
-    const decorative = this.props.decorative || false;
+  const decorative = props.decorative || false;
+  const sideCols = props.sideCols || (decorative ? 12 : 7);
+  const height = props.height || (decorative ? 2 : 7);
+  const width = props.width || 5;
 
-    const sideCols = this.props.sideCols || (decorative ? 12 : 7);
-    const height = this.props.height || (decorative ? 2 : 7);
-    const width = this.props.width || 5;
-    return (
-      <div className={`banner ${decorative ? 'decorative' : ''}`}>
-        <div className="square-col-container">
-          {!decorative &&
-            generateCols(1, height, [
-              '',
-              '',
-              '',
-              '',
-              'white',
-              'white',
-              'white',
-            ])}
-          {!decorative &&
-            generateCols(1, height, ['', '', '', '', '', 'white', 'white'])}
-          {generateCols(1, height, ['', '', '', '', '', '', 'white'])}
-          {generateCols(sideCols, height, undefined, this.state.randomize)}
-          {!decorative &&
-            generateCols(1, height, ['', '', 'white', 'white', '', '', ''])}
-          {!decorative &&
-            generateCols(width, height, [
-              '',
-              '',
-              'white',
-              'white',
-              'white',
-              '',
-              '',
-            ])}
-          {!decorative &&
-            generateCols(1, height, ['', '', '', 'white', 'white', '', ''])}
-          {generateCols(sideCols, height, undefined, this.state.randomize)}
-          {generateCols(1, height, ['white', '', '', '', '', '', ''])}
-          {!decorative &&
-            generateCols(1, height, ['white', 'white', '', '', '', '', ''])}
-          {!decorative &&
-            generateCols(1, height, [
-              'white',
-              'white',
-              'white',
-              '',
-              '',
-              '',
-              '',
-            ])}
-        </div>
-        {!decorative && (
-          <div className="title">
-            <h1>code the future.</h1>
-          </div>
-        )}
+  return (
+    <div className={`banner ${decorative ? 'decorative' : ''}`}>
+      <div className="square-col-container">
+        {!decorative &&
+          generateCols(1, height, [
+            '',
+            '',
+            '',
+            '',
+            'white',
+            'white',
+            'white',
+          ])}
+        {!decorative &&
+          generateCols(1, height, ['', '', '', '', '', 'white', 'white'])}
+        {generateCols(1, height, ['', '', '', '', '', '', 'white'])}
+        {generateCols(sideCols, height, undefined, randomize)}
+        {!decorative &&
+          generateCols(1, height, ['', '', 'white', 'white', '', '', ''])}
+        {!decorative &&
+          generateCols(width, height, [
+            '',
+            '',
+            'white',
+            'white',
+            'white',
+            '',
+            '',
+          ])}
+        {!decorative &&
+          generateCols(1, height, ['', '', '', 'white', 'white', '', ''])}
+        {generateCols(sideCols, height, undefined, randomize)}
+        {generateCols(1, height, ['white', '', '', '', '', '', ''])}
+        {!decorative &&
+          generateCols(1, height, ['white', 'white', '', '', '', '', ''])}
+        {!decorative &&
+          generateCols(1, height, [
+            'white',
+            'white',
+            'white',
+            '',
+            '',
+            '',
+            '',
+          ])}
       </div>
-    );
-  }
-}
+      {!decorative && (
+        <div className="title">
+          <h1>code the future.</h1>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default Banner;
