@@ -30,6 +30,7 @@ module.exports = {
     'eol-last': ['error', 'always'],
     'react/prop-types': 'off',
     'react/no-unknown-property': 'off',
+    'react/react-in-jsx-scope': 'off',
 
     // many of these rules are taken from our friends at Creative Labs;
     // see their config here: https://github.com/UCLA-Creative-Labs/sunshine/blob/master/.eslintrc.js
@@ -87,7 +88,7 @@ module.exports = {
     'quote-props': ['error', 'consistent-as-needed'],
 
     // No multiple empty lines
-    'no-multiple-empty-lines': ['error'],
+    'no-multiple-empty-lines': ['error', { max: 1 }],
 
     // Max line lengths
     'max-len': [
