@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { axe, toHaveNoViolations } from 'jest-axe';
-import React from 'react';
-import { act } from 'react-dom/test-utils';
+
+import { act } from 'react';
 import Footer from '../../components/Footer';
 
 expect.extend(toHaveNoViolations);
