@@ -1,5 +1,5 @@
 // import Image from 'next/image';
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 function SidebarLink(props) {
   return (
