@@ -1,5 +1,5 @@
 // import Image from 'next/image';
-import React from 'react';
+
 import CommitteeEventCard from './CommitteeEventCard';
 import Intro from './CommitteeSectionIntro';
 
