@@ -1,4 +1,4 @@
-import React from 'react';
+
 import Countdown from 'react-countdown';
 
 export default function CountdownElement(props) {
