@@ -1,7 +1,7 @@
 // import { faEnvelope } from '@fortawesome/free-regular-svg-icons';
 // import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import Image from 'next/legacy/image';
-import React from 'react';
+
 import styles from '../../styles/components/Officers/OfficerCard.module.scss';
 import '@fortawesome/fontawesome-svg-core/styles.css';
 
