@@ -1,4 +1,4 @@
-import React from 'react';
+
 import ProjectCard from './ProjectCard';
 
 function Projects(props) {
