@@ -1,6 +1,4 @@
 /* eslint-disable max-len */
-
-import React from 'react';
 import styles from '../styles/pages/Internship.module.scss';
 
 export default {
@@ -47,27 +45,34 @@ export default {
       name: 'Maya Raman',
       quote: (
         <>
-          <h3 className={styles['testimonial-question']}>
+          <p className={styles['testimonial-question']}>
             What is your favorite memory from being an intern?
-          </h3>
-          <span>Accidentally breaking tables at the intern retreat</span>
-          <h3 className={styles['testimonial-question']}>
+          </p>
+          <p className={styles['testimonial-answer']}>
+            Accidentally breaking tables at the intern retreat.
+          </p>
+
+          <p className={styles['testimonial-question']}>
             How much CS experience did you have before you became an intern?
-          </h3>
-          <span>
-            Not much - I took AP CS in high school and didn’t learn a ton, and
-            when I applied to be an intern I was taking CS 31.
-          </span>
-          <h3 className={styles['testimonial-question']}>
+          </p>
+          <p className={styles['testimonial-answer']}>
+            Not much - I took AP CS in high school and didn’t learn a ton, and when
+            I applied to be an intern I was taking CS 31.
+          </p>
+
+          <p className={styles['testimonial-question']}>
             Are you still an officer in ACM? If so, in what Committee?
-          </h3>
-          <span>Yes! I’m the director of ACM Impact!</span>
-          <h3 className={styles['testimonial-question']}>
+          </p>
+          <p className={styles['testimonial-answer']}>
+            Yes! I’m the director of ACM Impact!
+          </p>
+
+          <p className={styles['testimonial-question']}>
             What advice would you give to prospective interns?
-          </h3>
-          <span>
+          </p>
+          <p className={styles['testimonial-answer']}>
             Try to maximize the amount of rejections you get - try everything!
-          </span>
+          </p>
         </>
       ),
       committee: 'AI',
@@ -77,37 +82,40 @@ export default {
       name: 'Archisha Datta',
       quote: (
         <>
-          <h3 className={styles['testimonial-question']}>
+          <p className={styles['testimonial-question']}>
             What is your favorite memory from being an intern?
-          </h3>
-          <span>
+          </p>
+          <p className={styles['testimonial-answer']}>
             My favorite memory is the first meeting we had for the learning lab
             I was working on. At the time, it seemed really scary to be thrown
             into a project with not a lot of experience, but looking back on it
             I’m proud of how much I learned since then.
-          </span>
-          <h3 className={styles['testimonial-question']}>
+          </p>
+
+          <p className={styles['testimonial-question']}>
             How much CS experience did you have before you became an intern?
-          </h3>
-          <span>
+          </p>
+          <p className={styles['testimonial-answer']}>
             I only knew basic programming (CS31) level when I became an intern.
-          </span>
-          <h3 className={styles['testimonial-question']}>
+          </p>
+
+          <p className={styles['testimonial-question']}>
             Are you still an officer in ACM? If so, in what Committee?
-          </h3>
-          <span>
+          </p>
+          <p className={styles['testimonial-answer']}>
             Yes, I’m the design director for TeachLA and I’m also getting
             involved in Design starting fall!
-          </span>
-          <h3 className={styles['testimonial-question']}>
+          </p>
+
+          <p className={styles['testimonial-question']}>
             What advice would you give to prospective interns?
-          </h3>
-          <span>
+          </p>
+          <p className={styles['testimonial-answer']}>
             If you’re on the fence about applying because you think you don’t
             have enough experience or you don’t feel like you belong in CS
             culture, I would give it a shot anyways! You might surprise
             yourself.
-          </span>
+          </p>
         </>
       ),
       committee: 'Teach LA',
