@@ -1,6 +1,6 @@
 import Image from 'next/legacy/image';
 import { NextSeo } from 'next-seo';
-import React from 'react';
+
 import Banner from '../components/Banner';
 import CommitteeSection from '../components/Committees/CommitteeSection';
 import Navigation from '../components/Committees/Sidebar';
