@@ -1,6 +1,6 @@
 import moment from 'moment';
 import { NextSeo } from 'next-seo';
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Calendar, momentLocalizer } from 'react-big-calendar';
 
 import Banner from '../components/Banner';
