@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { NextSeo } from 'next-seo';
-import React from 'react';
+
 import Countdown from 'react-countdown';
 
 import Banner from '../../components/Banner';
@@ -123,7 +123,7 @@ function GMCountdown(props) {
         target="_blank"
         rel="noreferrer noopener">
 
-          RSVP Now!
+        RSVP Now!
 
       </Link>
     </div>
@@ -182,7 +182,7 @@ function F21() {
                 (learn more about how ACM is{' '}
                 <Link href={'/covid'} target="_blank" rel="noreferrer noopener">
 
-                    committed to running COVID-safe events
+                  committed to running COVID-safe events
 
                 </Link>{' '}
                 in our return to campus)
