@@ -1,7 +1,7 @@
 import Image from 'next/legacy/image';
 import Link from 'next/link';
 import { NextSeo } from 'next-seo';
-import React from 'react';
+
 import Countdown from 'react-countdown';
 
 import Banner from '../../components/Banner';
@@ -115,7 +115,7 @@ function GMCountdown(props) {
         target="_blank"
         rel="noreferrer noopener">
 
-          RSVP Now!
+        RSVP Now!
 
       </Link>
     </div>
