@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { NextSeo } from 'next-seo';
-import React from 'react';
+
 import Countdown from 'react-countdown';
 
 import Banner from '../../components/Banner';
@@ -107,7 +107,7 @@ function GMCountdown(props) {
         target="_blank"
         rel="noreferrer noopener">
 
-          RSVP Now!
+        RSVP Now!
 
       </Link>
     </div>
