@@ -1,10 +1,7 @@
 import Image from 'next/legacy/image';
 import Link from 'next/link';
 import { NextSeo } from 'next-seo';
-import React from 'react';
-
 import Layout from '../../components/Layout';
-
 import { blogLink, contactEmail } from '../../data/jedi';
 import diverseDesign from '../../public/images/jedi/events/diverse-design.png';
 import hackschoolAccessibility from '../../public/images/jedi/events/hackschool-accessibility.png';
@@ -60,7 +57,7 @@ function JEDIAbout() {
               <li>
                 <Link href={blogLink} target="_blank" rel="noopener noreferrer">
 
-                    blog
+                  blog
 
                 </Link>
               </li>
@@ -87,7 +84,7 @@ function JEDIAbout() {
                 target="_blank"
                 rel="noopener noreferrer">
 
-                  in this blog post
+                in this blog post
 
               </Link>
               . We also hold allyship spaces on various EDI topics (not always
@@ -98,7 +95,7 @@ function JEDIAbout() {
                 target="_blank"
                 rel="noopener noreferrer">
 
-                  queer and trans representation in computer science
+                queer and trans representation in computer science
 
               </Link>{' '}
               and{' '}
@@ -108,7 +105,7 @@ function JEDIAbout() {
                 target="_blank"
                 rel="noopener noreferrer">
 
-                  Asian representation and the otherization of kpop
+                Asian representation and the otherization of kpop
 
               </Link>
               .
@@ -128,7 +125,7 @@ function JEDIAbout() {
                 target="_blank"
                 rel="noopener noreferrer">
 
-                  read our blog
+                read our blog
 
               </Link>
               !
